Before debugbar

+++ app/Http/Controllers/PostsController.php
@@ -4,11 +4,18 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 // use Illuminate\Support\Facades\DB;
 
 class PostsController extends Controller
 {
+
+  public function __construct() {
+    $this->middleware('auth')
+      ->only(['create', 'store', 'edit', 'update', 'destroy']);
+  }
     /* php artisan make:controller PostsController --resource
     As a resource controller I am made with CRUD routes in place. */ 
 
@@ -29,7 +36,11 @@ class PostsController extends Controller
       // dd(DB::getQueryLog());
       return view(
         'posts.index', 
-        ['posts' => BlogPost::withCount('comments')->get()]
+        ['posts' => BlogPost::latest()->withCount('comments')->get(),
+          'mostCommented' => BlogPost::mostCommented()->take(5)->get(),
+          'mostActive' => User::withMostBlogPosts()->take(5)->get(),
+          'mostActiveLastMonth' => User::withMostBlogPostsLastMonth()->take(5)->get()
+        ]
       );
     }
 
@@ -40,6 +51,7 @@ class PostsController extends Controller
      */
     public function create()
     {
+      // $this->authorize('posts.create');
       return view('posts.create');
     }
 
@@ -52,6 +64,7 @@ class PostsController extends Controller
     public function store(StorePost $request)
     {
       $validated = $request->validated();
+      $validated['user_id'] = $request->user()->id;
       $post = BlogPost::create($validated);
 
       $request->session()->flash('status', 'Post created');
@@ -79,6 +92,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
+      $post = BlogPost::findOrFail($id);
+      // if (Gate::denies('update-post', $post)) {
+      //   abort(403, 'Nacho post! [Edit]');
+      // };
+      $this->authorize($post);
       return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
     }
 
@@ -92,6 +110,10 @@ class PostsController extends Controller
     public function update(StorePost $request, $id)
     {
       $post = BlogPost::findOrFail($id);
+      // if (Gate::denies('update-post', $post)) {
+      //   abort(403, 'Nacho post!');
+      // };
+      $this->authorize($post);
       $validated = $request->validated();
       $post->fill($validated)->save();
       $request->session()->flash('status', 'Post updated');
@@ -107,6 +129,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
       $post = BlogPost::findOrFail($id);
+      $this->authorize($post);
+      // if (Gate::denies('delete-post', $post)) {
+      //   abort(403, 'Nacho post! [Delete]');
+      // };
       $post->delete();
 
       session()->flash('status', 'Post deleted');
